高速化・取得日の追加
