+++ scanner/Data/Entities/BaseEntity.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace scanner.Data.Entities
+{
+        internal interface BaseEntity
+        {
+            int id { get; set; }
+        }
+
+}
